+++ spec/features/analytics_features_spec.rb
@@ -7,7 +7,8 @@ RSpec.describe 'User interacts with Analytics' do
     before :each do
       @chapter = create :chapter
       @group = create :group, chapter: @chapter
-      create :graded_student, group: @group, grades: { 'Memorization' => [3, 4, 5, 6, 7], 'Grit' => [2, 3, 2, 4, 5] }
+      @student = create :graded_student, group: @group, grades: { 'Memorization' => [3, 4, 5, 6, 7], 'Grit' => [2, 3, 2, 4, 5] }
+      create :enrollment, group: @group, student: @student, active_since: 1.year.ago
       @organization = @chapter.organization
       Scenic.database.refresh_materialized_view('student_lesson_summaries')
     end